refactor: rename `SetLikeExtension` -> `SetExtension` and `MappingStreams` -> `FunctionSpaces`
